+++ packages/image_picker/image_picker_for_web/test/image_picker_for_web_test.dart
@@ -12,9 +12,10 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:image_picker_for_web/image_picker_for_web.dart';
 import 'package:image_picker_platform_interface/image_picker_platform_interface.dart';
 
-final String expectedStringContents = "Hello, world!";
+final String expectedStringContents = 'Hello, world!';
+final expectedSize = expectedStringContents.length;
 final Uint8List bytes = utf8.encode(expectedStringContents);
-final html.File textFile = html.File([bytes], "hello.txt");
+final html.File textFile = html.File([bytes], 'hello.txt');
 
 void main() {
   // Under test...
@@ -34,7 +35,7 @@ void main() {
     final plugin = ImagePickerPlugin(overrides: overrides);
 
     // Init the pick file dialog...
-    final file = plugin.pickFile();
+    final file = plugin.pickFileFromBrowser();
 
     // Mock the browser behavior of selecting a file...
     mockInput.dispatchEvent(html.Event('change'));
@@ -42,10 +43,46 @@ void main() {
     // Now the file should be available
     expect(file, completes);
     // And readable
-    expect((await file).readAsBytes(), completion(isNotEmpty));
+    final pickedFile = await file;
+    expect(pickedFile.readAsBytes(), completion(isNotEmpty));
+    expect(pickedFile.length(), completion(equals(expectedSize)));
+    expect(pickedFile.name, 'hello.txt');
   });
 
   // There's no good way of detecting when the user has "aborted" the selection.
+
+  test('computeAcceptAttribute', () {
+    expect(
+      plugin.computeAcceptAttribute(null),
+      '',
+    );
+    expect(
+      plugin.computeAcceptAttribute([]),
+      '',
+    );
+    expect(
+      plugin.computeAcceptAttribute([null, '', '  ']),
+      '',
+      reason: 'Should remove null/empty values and end up with an empty string',
+    );
+    expect(
+      plugin.computeAcceptAttribute(['jpg', 'png', 'bmp']),
+      '.jpg,.png,.bmp',
+    );
+    expect(
+      plugin.computeAcceptAttribute(['.jpg', '.png', '.bmp']),
+      '.jpg,.png,.bmp',
+    );
+    expect(
+      plugin.computeAcceptAttribute(['.jpg', 'png', '.bmp']),
+      '.jpg,.png,.bmp',
+    );
+    expect(
+      plugin.computeAcceptAttribute([null, '.jpg', '', 'png', '  ', '.bmp']),
+      '.jpg,.png,.bmp',
+      reason: 'Should strip out null/empty values from the list of extensions',
+    );
+  });
 
   test('computeCaptureAttribute', () {
     expect(
@@ -67,6 +104,12 @@ void main() {
   });
 
   group('createInputElement', () {
+    test('accept: empty string', () {
+      html.Element input = plugin.createInputElement('', null);
+
+      expect(input.attributes, isNot(contains('accept')));
+    });
+
     test('accept: any, capture: null', () {
       html.Element input = plugin.createInputElement('any', null);