+++ packages/image_picker/image_picker_for_web/CHANGELOG.md
@@ -1,7 +1,12 @@
-# 0.1.0+1
+## 0.1.1
+
+* Implement `pickArbitraryFile` platform method.
+* Initialize `name` and `size` of the picked file.
+
+## 0.1.0+1
 
 * Remove `android` directory.
 
-# 0.1.0
+## 0.1.0
 
 * Initial open-source release.